requirements.txt
cloud-sql-python-connector[pg8000] >= 1.4.1
functions-framework==3.*
python-dotenv==1.0.0